requirements.txt
# Core dependencies
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
transformers==4.35.2
indic-nlp-library==0.92
sentencepiece==0.1.99
protobuf==4.25.3

# API dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
cachetools==5.3.2
httpx==0.25.2
python-multipart==0.0.6

# Monitoring and logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0 